+++ Core/TgInfrastructure/License/TgLicense.cs
@@ -4,23 +4,54 @@
 namespace TgInfrastructure.License;
 
 [DebuggerDisplay("{ToDebugString()}")]
-public abstract partial class TgLicense : ObservableRecipient
+public partial class TgLicense : ObservableRecipient
 {
 	#region Public and private fields, properties, constructor
 
 	[ObservableProperty]
-	public partial string Description { get; set; } = "Empty license";
+	public partial string Description { get; private set; } = "Empty license";
 	[ObservableProperty]
-	public partial string LicenseKey { get; set; } = string.Empty;
-	public TgEnumLicenseType LicenseType { get; set; } = TgEnumLicenseType.Free;
+	public partial Guid LicenseKey { get; set; } = Guid.Empty;
+	[ObservableProperty]
+	public partial TgEnumLicenseType LicenseType { get; set; } = TgEnumLicenseType.Free;
+	[ObservableProperty]
+	public partial long UserId { get; set; }
+	[ObservableProperty]
+	public partial DateTime ValidTo { get; set; } = DateTime.MinValue;
+	[ObservableProperty]
+	public partial bool IsConfirmed { get; set; }
 
 	#endregion
 
 	#region Public and private methods
 
-	public virtual bool IsValid() => false;
+	public bool IsValid()
+	{
+		switch (LicenseType)
+		{
+			case TgEnumLicenseType.Free:
+				return true;
+			case TgEnumLicenseType.Test:
+			case TgEnumLicenseType.Paid:
+				return IsConfirmed && DateTime.Now < ValidTo;
+			case TgEnumLicenseType.Premium:
+				break;
+		}
+		return false;
+	}
 
 	public string ToDebugString() => Description;
+
+	public void SetDescription(string description)
+	{
+		Description = description;
+	}
+
+	public string GetLicenseKeyString() => LicenseKey == Guid.Empty ? "-" : $"{LicenseKey:D}";
+
+	public string GetUserIdString() => UserId == 0 ? "-" : $"{UserId}";
+
+	public string GetValidToString() => ValidTo <= DateTime.MinValue ? "-" : $"{ValidTo:yyyy-MM-dd}";
 
 	#endregion
 }