+++ Docs/CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.6.150] - 2025-05-06
+## [0.6.160] - 2025-05-07
 ### Changed
 - Updated the menu in TgDownloaderConsole
+- Updated license view in TgDownloaderConsole
+- Updated license view in OpenTgResearcherDesktop
 ### Added
 - Added methods to memorize window size and position between sessions in OpenTgResearcherDesktop